Add handling for checking if the ot3 firmware repo has changed
The ADD hits the github api which returns a comparision of main and HEAD. 
If ot3-firmware has changed then the returned value will be different than the cached value. This will break the cache at this step and all the subsequent steps will be ran again.
This will re-clone the git repo so we don't have to worry it changing out from under us and us forgetting to run a Docker build with --no-cache