chore: Cleanup code generation stuff (#51)
When using sed to add type ignores to compose_file_model.py, if lines are wrapped then it's super hard to add the type ignore. 

The package used to generate compose file model.py, datamodel-code-generator, uses black under the hood to format the generated code, which wraps the lines.

What needed to happen is the line length needed to be increased for only compose_file_model.py. 

To do this, I set the line-length in pyproject.toml to 240. datamodel-code-generator will read from this toml file. 
To avoid having all the other files having their line lengths set to 240, I added the --line-length option to the black calls in the Makefile.