Rework container filters
Needed to be able to specify hardware_only or firmware_only in the filter itself.
This is required to get a list of container names to pip install monorepo packages on. 
Need to get the `firmware_only` modules that run off of the  monorepo and not hardware_only that runs off of opentrons-modules
Fix names and actually have correct syntax