Enforce startup order of containers. Remove some spammy logging
With this massive disgusting refactor, one of the things that cannot be enforced with docker, is container startup order. 

Previously all containers were started at the same time. The issue with this was 2 fold: 

- ot3-firmware emulators would start before the can-server, timeout, and die
- module emulators would start before the emulator proxy, timeout, and die

To fix this, the startup control is now being given to the Makefile. The flow is as follows:

- The emulator proxy and can server are started and given a little bit of time to come up (2 seconds). These are ran in the background
- Firmware (smoothie for OT-2, ot3-firmware for Flex) and modules are started. These are ran in the background
- The robot server is started. This is started in the foreground. 

Pushing commands to the background removes spammy logging like G-Code stuff